--- src/app/[auth]/profile/profile.css
@@ -1,31 +1,0 @@
-.content {
-  width: calc(100% - 600px);
-}
-
-.content_1 {
-  width: calc(100% - 480px);
-}
-
-.item {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 6px;
-  flex: 1 0 0;
-}
-
-.row {
-  display: flex;
-  align-items: flex-start;
-  gap: 24px;
-}
-
-.footer {
-  bottom: 8px;
-  position: fixed;
-  justify-content: flex-end;
-}
-
-.profileBanner {
-  background-image: url(https://storage.googleapis.com/sorbet_profile/banner65967ccea5b2ae53bc4ccfa61705131804225.png);
-}
