+++ src/components/profile/index.ts
@@ -5,5 +5,8 @@ import { WidgetContainer } from "./widget-container";
 import { ResizeWidget } from "./resize-widget"; 
 import { AddWidgets } from "./add-widgets"; 
 import { InvalidAlert } from "./invalid-alert"; 
+import { ProfileEditModal } from "./profile-edit-modal";
+import { InputLocation } from "./input-location"; 
+import { InputSkills } from "./input-skills"; 
 
-export { Profile, ProfileHeader, Widget, WidgetContainer, ResizeWidget, AddWidgets, InvalidAlert };
+export { Profile, ProfileHeader, Widget, WidgetContainer, ResizeWidget, AddWidgets, InvalidAlert, ProfileEditModal, InputLocation, InputSkills };
