+++ src/components/profile/input-location.tsx
@@ -1,17 +1,19 @@
 /* eslint-disable @next/next/no-img-element */
-import { Dispatch, SetStateAction, useEffect, useState } from 'react';
-
-import Autocomplete from 'react-google-autocomplete';
-import UserType from '@/types/user';
 import { config } from '@/lib/config';
+import { cn } from '@/lib/utils';
+import type { User } from '@/types';
+import { MapPin } from 'lucide-react';
+import { Dispatch, SetStateAction, useEffect, useState } from 'react';
+import Autocomplete from 'react-google-autocomplete';
 
 interface LocationProps {
-  userData: UserType;
-  setUserData: Dispatch<SetStateAction<UserType>>;
+  userData: User;
+  setUserData: Dispatch<SetStateAction<User>>;
   onInputChange: (e: any) => void;
   defaultValue: string | undefined;
 }
-const Location: React.FC<LocationProps> = ({
+
+export const InputLocation: React.FC<LocationProps> = ({
   userData,
   setUserData,
   onInputChange,
@@ -28,30 +30,27 @@ const Location: React.FC<LocationProps> = ({
   }, [place]);
 
   return (
-    <div className='h-full w-full'>
-      <div className='relative h-full w-full items-center justify-center rounded-md outline-none'>
-        <img
-          src='/svg/location.svg'
-          alt='location'
-          width={20}
-          height={20}
-          className='absolute left-[14px] top-[10px]'
-        />
+    <div className='flex flex-row gap-4 bg-white w-full mt-1'>
+      <div
+        className={cn(
+          'flex items-center border border-2 border-gray-200 px-2 rounded-xl flex-grow '
+        )}
+      >
+        <div>
+          <MapPin stroke={`gray`} />
+        </div>
         <Autocomplete
           apiKey={config.googleMapKey}
           onPlaceSelected={(place) => setPlace(place)}
           onChange={(e) => onInputChange(e)}
           defaultValue={defaultValue}
-          className={`h-full w-full rounded-md
-          border-[#D0D5DD] p-[10px] pl-[42px] text-base font-normal text-[#667085]`}
+          className={`h-full w-full rounded-md border-[#D0D5DD] p-[10px] text-base font-normal text-[#667085] outline-none`}
           options={{
             types: ['(cities)'],
           }}
-          placeholder='Search location'
+          placeholder='Enter your city, country'
         />
       </div>
     </div>
   );
 };
-
-export default Location;
