+++ src/types/user.ts
@@ -1,4 +1,4 @@
-export default interface UserType {
+export interface User {
   id: string;
   firstName: string;
   lastName: string;
